+++ src/main/java/frc/robot/constants/IdConstants.java
@@ -35,5 +35,12 @@ public class IdConstants {
     public static final int CLIMB_MOTOR = 20;
 
     // Outtake
-    public static final int OUTTAKE_MOTOR = -1;
+    /** CAN ID for the Alphabot's outake */
+    public static final int OUTTAKE_MOTOR_ALPHA = 14;
+    /** CAN ID for the competition bot's outtake */
+    public static final int OUTTAKE_MOTOR_COMP = -1; 
+
+    // Digital inputs
+    public static final int OUTTAKE_DIO_LOADED = 9;
+    public static final int OUTTAKE_DIO_EJECTING = 8;
 }
