+++ src/main/java/frc/robot/subsystems/Outtake.java
@@ -17,87 +17,14 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.constants.IdConstants;
 
-
-public class Outtake extends SubsystemBase {
-
-    private SparkFlex  motor = new SparkFlex(IdConstants.OUTTAKE_MOTOR, MotorType.kBrushless);
-    private double power;
-
-
-    /** Coral detected before the rollers */
-    private DigitalInput digitalInputLoaded = new DigitalInput(9);
-    private DIOSim dioInputLoaded;
-    /** Coral detected after the rollers */
-    private DigitalInput digitalInputEjecting = new DigitalInput(8);
-    private DIOSim dioInputEjecting;
-    private int ticks = 0;
-
-
-    public Outtake(){
-        motor.configure(new SparkFlexConfig()
-            .inverted(true)
-            .idleMode(IdleMode.kBrake),
-            ResetMode.kResetSafeParameters,
-            PersistMode.kNoPersistParameters
-        );
-        if (RobotBase.isSimulation()){
-            // object that will control the loaded sensor
-            dioInputLoaded = new DIOSim(digitalInputLoaded);
-            // object that will control the ejecting sensor
-            dioInputEjecting = new DIOSim(digitalInputEjecting);
-            // assume coral is loaded
-            dioInputLoaded.setValue(false);
-            // we are not ejecting
-            dioInputEjecting.setValue(true);
-        }
-    }
-
-
-    @Override
-    public void periodic(){
-        motor.set(power);
-        SmartDashboard.putBoolean("Coral loaded", coralLoaded());
-        SmartDashboard.putBoolean("Coral ejected", coralEjecting());
-    }
-
-
-    @Override
-    public void simulationPeriodic(){
-        // when coral is ejecting, loading is true & ejecting is true. when coral shoots out, loading is false & ejecting is false
-        ticks++;
-
-
-        if (motor.get() > 0.05) {
-            if (ticks > 250) {
-                ticks = 0;
-            }
-            // motor is outtaking
-            // motor is spinning, ejecting will be true. after 0.14 seconds
-            if (ticks ==7) {
-                dioInputEjecting.setValue(false);
-            }
-            if (ticks == 14){
-                // after 0.14 seconds
-                dioInputLoaded.setValue(true);
-            }
-            if (ticks == 16){
-                // after 0.18 seconds
-                dioInputEjecting.setValue(true);
-            }
-        }
-
-
-        if (ticks == 250) {
-            // make coral appear again (set to true)
-            dioInputLoaded.setValue(false);
-        }
-    }
+/**
+ * Abstract class for the outtake. All commands should use this subsystem
+ */
+public abstract class Outtake extends SubsystemBase {
 
 
     /** Set the motor power to move the coral */
-    public void setMotor(double power){
-        this.power = power;
-    }
+    public abstract void setMotor(double power);
 
 
     /** stop the coral motor */
@@ -115,26 +42,17 @@ public class Outtake extends SubsystemBase {
     }
 
 
-    public boolean coralLoaded(){
-       return !digitalInputLoaded.get();
-    }
+    public abstract boolean coralLoaded();
 
 
     /**
      *  Coral is at the ejecting beam break sensor.
      * @return coral is interrupting the beam breaker.
      */
-    public boolean coralEjecting() {
-        return !digitalInputEjecting.get();
-    }
+    public abstract boolean coralEjecting();
 
 
     public void reverse(){
         setMotor(-0.2);
-    }
-
-
-    public boolean isSimulation(){
-        return RobotBase.isSimulation();
     }
 }
